test-requirements.txt
nose2
flask-testing
Werkzeug
requests